+++ internal/domain/book/request.go
@@ -8,7 +8,7 @@ type CreateRequest struct {
 }
 
 type UpdateRequest struct {
-	ID            int    `json:"-"`
+	ID            uint64 `json:"-"`
 	Title         string `json:"title" validate:"required"`
 	PublishedDate string `json:"published_date" validate:"required"`
 	ImageURL      string `json:"image_url" validate:"url"`
